+++ expyriment/misc/_clock.py
@@ -12,11 +12,14 @@ __version__ = ''
 __revision__ = ''
 __date__ = ''
 
+
 import sys
 import time
 from types import FunctionType
+
 from ._timer import get_time
 from .. import _internals
+
 
 class Clock(object) :
     """Basic timing class.
@@ -42,7 +45,6 @@ class Clock(object) :
 
         """
 
-        from ..io.defaults import _skip_wait_functions
         if (sync_clock.__class__.__name__ == "Clock"):
             self.__init_time = sync_clock.init_time // 1000
         else:
@@ -50,8 +52,6 @@ class Clock(object) :
 
         self._init_localtime = time.localtime()
         self.__start = get_time()
-
-        self._skip_wait_functions = _skip_wait_functions
 
 
     @staticmethod
@@ -103,8 +103,8 @@ class Clock(object) :
 
         self.__start = get_time()
 
-    def wait(self, waiting_time, function=None):
-        """Wait for a certain amout of milliseconds.
+    def wait(self, waiting_time, callback_function=None, process_control_events=False):
+        """Wait for a certain amount of milliseconds.
 
         Parameters
         ----------
@@ -112,78 +112,111 @@ class Clock(object) :
             time to wait in milliseconds
         function : function, optional
             function to repeatedly execute during waiting loop
+        process_control_events : bool, optional
+            process ``io.Keyboard.process_control_keys()`` and
+            ``io.Mouse.process_quit_event()`` (default = False)
 
         Returns
         -------
         quit_event : expyriment.control.CallbackQuitEvent object
            the callback quit even in case a wait function has been registered
-            
+
         See Also
         --------
         design.experiment.register_wait_callback_function
 
         """
 
-        if self._skip_wait_functions:
+        if _internals.skip_wait_functions:
             return
         start = self.time
-        if isinstance(function, FunctionType) or\
-                            _internals.active_exp.is_callback_registered:
+        if isinstance(callback_function, FunctionType) or \
+           (_internals.active_exp is not None and \
+            (process_control_events or \
+             _internals.active_exp.is_callback_registered)):
             while (self.time < start + waiting_time):
-                if isinstance(function, FunctionType):
-                    function()
-                rtn_callback = _internals.active_exp._execute_wait_callback()
-                if isinstance(rtn_callback, _internals.CallbackQuitEvent):
-                    return rtn_callback
+                if isinstance(callback_function, FunctionType):
+                    callback_function()
+                if _internals.active_exp.is_initialized:
+                    rtn_callback = _internals.active_exp._execute_wait_callback()
+                    if isinstance(rtn_callback, _internals.CallbackQuitEvent):
+                        return rtn_callback
+                    if process_control_events:
+                        if _internals.active_exp.keyboard.process_control_keys():
+                            break
+                        _internals.active_exp.mouse.process_quit_event()
+                    else:
+                        import pygame
+                        pygame.event.pump()
         else:
             looptime = 200
             if (waiting_time > looptime):
-                time.sleep((waiting_time - looptime) // 1000)
+                if _internals.active_exp is not None and \
+                   _internals.active_exp.is_initialized:
+                    while (self.time < start + (waiting_time - looptime)):
+                        if process_control_events:
+                            if _internals.active_exp.mouse.process_quit_event() or \
+                               _internals.active_exp.keyboard.process_control_keys():
+                                break
+                        else:
+                            import pygame
+                            pygame.event.pump()
+                else:
+                    time.sleep((waiting_time - looptime) // 1000)
             while (self.time < start + waiting_time):
                 pass
 
-    def wait_seconds(self, time_sec, function=None):
+    def wait_seconds(self, time_sec, callback_function=None,
+                     process_control_events=False):
         """Wait for a certain amout of seconds.
 
         Parameters
         ----------
         time_sec : int
             time to wait in seconds
-        function : function, optional
+        callback_function : function, optional
             function to repeatedly execute during waiting loop
+        process_control_events : bool, optional
+            process ``io.Keyboard.process_control_keys()`` and
+            ``io.Mouse.process_quit_event()`` (default = False)
 
         Returns
         -------
         quit_event : expyriment.control.CallbackQuitEvent object
            the callback quit even in case a wait function has been registered
-        
+
         See Also
         --------
         Clock.wait, design.experiment.register_wait_callback_function
 
         """
 
-        return self.wait(time_sec * 1000, function)
+        return self.wait(time_sec * 1000, callback_function, process_control_events)
 
-    def wait_minutes(self, time_minutes, function=None):
+    def wait_minutes(self, time_minutes, callback_function=None,
+                     process_control_events=False):
         """Wait for a certain amount of minutes.
 
         Parameters
         ----------
         time_minutes : int
             time to wait in minutes
-        function : function, optional
+        callback_function : function, optional
             function to repeatedly execute during waiting loop
+        process_control_events : bool, optional
+            process ``io.Keyboard.process_control_keys()`` and
+            ``io.Mouse.process_quit_event()`` (default = False)
 
         Returns
         -------
         quit_event : expyriment.control.CallbackQuitEvent object
            the callback quit even in case a wait function has been registered
-            
+
         See Also
         --------
         Clock.wait, design.experiment.register_wait_callback_function
 
         """
 
-        return self.wait_seconds(time_minutes * 60, function)
+        return self.wait_seconds(time_minutes * 60, callback_function,
+                                 process_control_events)